[mod] separate logic in imagescan
